Extract dagger modules to di package
