+++ app/src/main/java/gq/kirmanak/mealient/di/DisclaimerModule.kt
@@ -1,4 +1,4 @@
-package gq.kirmanak.mealient.ui.disclaimer
+package gq.kirmanak.mealient.di
 
 import dagger.Binds
 import dagger.Module
