+++ app/src/main/java/gq/kirmanak/mealient/di/UiModule.kt
@@ -1,4 +1,4 @@
-package gq.kirmanak.mealient.ui
+package gq.kirmanak.mealient.di
 
 import com.squareup.picasso.Picasso
 import dagger.Binds
@@ -6,6 +6,7 @@ import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
 import dagger.hilt.components.SingletonComponent
+import gq.kirmanak.mealient.ui.ImageLoader
 import gq.kirmanak.mealient.ui.picasso.ImageLoaderPicasso
 import gq.kirmanak.mealient.ui.picasso.PicassoBuilder
 import javax.inject.Singleton
